move HostFuncTrampoline::new constructor to where it belongs
